Extract ProjectStorageManager to an interface to be inherit-friendly
